% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwl-import.R
\name{addIdNum}
\alias{addIdNum}
\title{add \code{#} prefix to id}
\usage{
addIdNum(x)
}
\arguments{
\item{x}{(character) with \code{#} or not.}
}
\value{
a character with \code{#} prefix.
}
\description{
add \code{#} prefix to id
}
\examples{
addIdNum("bam")
}

