% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-app.R
\docType{class}
\name{App-class}
\alias{App}
\alias{App-class}
\title{App class}
\value{
App object.
}
\description{
App class
}
\section{Fields}{

\describe{
\item{\code{id}}{app id.}

\item{\code{project}}{project id}

\item{\code{name}}{app name}

\item{\code{revision}}{app revision}

\item{\code{raw}}{raw cwl list, if doesn't have any, call cwl() method.}
}}
\examples{
\dontrun{
a = Auth(url = "https://api.sbgenomics.com/v2/",
         token = "fake_token")
## get a flow
app = a$public_app(id = "admin/sbg-public-data/rna-seq-alignment-star")
app$input_matrix()
app$output_matrix()
## get a flow
app = a$public_app(id = "admin/sbg-public-data/star")
app$input_matrix()
app$output_matrix()
}
}

