% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cwl.R
\docType{class}
\name{CommandOutputParameter-class}
\alias{CommandOutputParameter-class}
\alias{CommandOutputParameter}
\title{CommandOutputParameter Class}
\description{
CommandOutputParameter Class
}
\section{Fields}{

\describe{
\item{\code{outputBinding}}{[CommandOutputBinding] Describes how to handle
the concrete outputs of a process step (such as files created by a
program) and describe them in the process output parameter.}
}}


\examples{
CommandOutputParameter(outputBinding = CommandOutputBinding(glob = "*.bam"))
}
