% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App-class.R
\name{convertApp}
\alias{appType}
\alias{convertApp}
\title{Convert App or a cwl JSON file to Tool or Flow object}
\usage{
convertApp(from)

appType(x)
}
\arguments{
\item{from}{an App object or a cwl JSON}

\item{x}{a App object}
}
\description{
Convert App or a cwl JSON file to Tool or Flow object
}
\details{
This function import cwl JSON file, based on its class: CommandLineTool or Worklfow
to relevant object in R, Tool object or Flow object.
}
\section{appType}{

\describe{
 this function return class of a App object.
}
}
\examples{
tool.in = system.file("extdata/app", "tool_star.json", package = "sevenbridges")
flow.in = system.file("extdata/app", "flow_star.json", package = "sevenbridges")
## convert to Tool object
convertApp(tool.in)
## convert to Flow object
convertApp(flow.in)
}

