% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cwl.R
\docType{class}
\name{CWL-class}
\alias{CWL-class}
\alias{CWL}
\title{Class CWL}
\description{
Define CWL class and generic methods, no fields defeind.
}
\section{Methods}{

\describe{
\item{\code{getFields(values)}}{Return fields as a list, used for following conversion, does not assume the value is a primitive type.}

\item{\code{toJSON(...)}}{Covert object to JSON}

\item{\code{toList(...)}}{Convert object to a list of simple data types}

\item{\code{toYAML(...)}}{Covert object to YAML}
}}

\examples{
# no fields, only to provide methods to be extended
x <- CWL()
}
