% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File-class.R, R/Task-class.R
\docType{methods}
\name{delete}
\alias{delete}
\alias{delete,Files-method}
\alias{delete,FilesList-method}
\alias{delete,SimpleList-method}
\alias{delete,Task-method}
\alias{download,Files-method}
\alias{download,FilesList-method}
\title{Delete file or files}
\usage{
delete(obj)

\S4method{delete}{SimpleList}(obj)

\S4method{delete}{Files}(obj)

\S4method{download}{FilesList}(obj, ...)

\S4method{download}{Files}(obj, ...)

\S4method{delete}{Task}(obj)
}
\arguments{
\item{obj}{single File or FileList}
}
\description{
Delete file
}
\examples{
\dontrun{
a$project("demo")$file("omni")$delete()
## or
delete(a$project("demo")$file("omni"))
}
}

