% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-files.R
\docType{methods}
\name{download}
\alias{download}
\alias{download,Files-method}
\alias{download,FilesList-method}
\title{Download files}
\usage{
download(obj, ...)

\S4method{download}{FilesList}(obj, ...)

\S4method{download}{Files}(obj, ...)
}
\arguments{
\item{obj}{single File or FileList}

\item{...}{passed to download()}
}
\value{
system message
}
\description{
Download files
}
\examples{
\dontrun{
a$project("demo")$file("omni")$download()
# or
download(a$project("demo")$file("omni"))}
}

