% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cwl.R
\name{DSCList}
\alias{DSCList}
\alias{DSCList-class}
\title{DSC list}
\usage{
DSCList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\value{
a DSCList
}
\description{
Contains DataypeSingleEnum, Schema, character
}
\examples{
DSCList("test", DatatypeEnum(), Schema())
}
