% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cwl.R
\docType{class}
\name{PrimitiveSingleEnum-class}
\alias{PrimitiveSingleEnum-class}
\alias{PrimitiveEnum}
\alias{ComplexSingleEnum-class}
\alias{ComplexEnum}
\alias{DatatypeSingleEnum-class}
\alias{DatatypeEnum}
\alias{enum-class}
\alias{enum}
\alias{ItemArray-class}
\alias{ItemArray}
\title{Pre-defiend enums}
\description{
Please check \code{cwl:::.CWL.Pritimive}, \code{cwl:::.CWL.Complex}.
}

\examples{
PrimitiveEnum()
PrimitiveEnum("boolean")
ComplexEnum("record")
DatatypeEnum("map")
}
