% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cwl.R
\docType{class}
\name{Expression-class}
\alias{Expression}
\alias{Expression-class}
\title{Expression Class}
\description{
Define an expression that will be evaluated and used to modify the
behavior of a tool or workflow. See Expressions for more
information about expressions and ExpressionEngineRequirement for
information on how to define a expression engine.
}
\section{Fields}{

\describe{
\item{\code{engine}}{(JsonPointerORcharacter) Either cwl:JsonPointer or a
reference to an ExpressionEngineRequirement defining which engine
to use.}

\item{\code{script}}{(character) The code to be executed by the expression
engine.}
}}
\examples{
Expression(engine = "#cwl-js-engine", script = "$job.inputs['threads']")
}

