% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cwl.R
\docType{class}
\name{ExpressionTool-class}
\alias{ExpressionTool-class}
\alias{ExpressionTool}
\title{ExpressionTool Class}
\description{
Execute an expression as a process step.
}
\section{Fields}{

\describe{
\item{\code{expression}}{(Expression) The expression to execute. The
expression must return a JSON object which matches the output
parameters of the ExpressionTool.}
}}


\examples{
ExpressionTool(
  expression =
    Expression(
      engine = "cwl:JsonPointer",
      script = "$job.inputs['threads']"
    )
)
}
