% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{file_copy}
\alias{file_copy}
\title{Copy specified file(s) to the specified project}
\usage{
file_copy(auth_token = NULL, project_id = NULL, file_id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to copy files to.}

\item{file_id}{Character vector. IDs of the files you wish to copy to
the project.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Copy specified file(s) to the specified project
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = file_copy(token,
            project_id = '7f7a72d0-da77-4f51-9416-99f14f7316ab',
            file_id = c('5506a44ae4b04a4ab3ae7250',
                        '5506a44ae4b04a4ab3ae7254',
                        '5506a44ae4b04a4ab3ae7252'))}
}

