% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{file_delete}
\alias{file_delete}
\title{Removes a file from a project}
\usage{
file_delete(auth_token = NULL, project_id = NULL, file_id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{file_id}{ID of a file you want to delete.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Removes a file from a project
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = file_delete(token,
            project_id = '1c1d06d2-5862-48f6-b595-e0099b20937e',
            file_id = '530854e2e4b036506b803c7e')}
}

