% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{file_download_url}
\alias{file_download_url}
\title{Returns a direct download URL for a project's file}
\usage{
file_download_url(auth_token = NULL, project_id = NULL, file_id = NULL,
  ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{file_id}{ID of a file you want to access.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns a direct download URL for a project's file.
}
\details{
You can use any HTTP client, or library to access or download
the content once you get the URL.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = file_download_url(token,
            project_id = '1c1d06d2-5862-48f6-b595-e0099b20937e',
            file_id = '530854e2e4b036506b803c7e')}
}

