% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{file_list}
\alias{file_list}
\title{Returns the list of all project files for a project}
\usage{
file_list(auth_token = NULL, project_id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.
Note that specifying \code{"public"} you can list public files.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns the list of all project files for a project. If user specifies
string \code{"public"} as \code{project_id}, this will return a list
of public files.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = file_list(token,
            project_id = '1c1d06d2-5862-48f6-b595-e0099b20937e')}
}

