% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File-class.R
\docType{class}
\name{Files-class}
\alias{Files}
\alias{Files-class}
\alias{FilesList}
\alias{FilesList-class}
\title{Files class}
\usage{
FilesList(...)
}
\arguments{
\item{...}{one or more Files objects}
}
\value{
Files object
}
\description{
Files Class
}
\details{
Files (with "s") class contains CWL File class, it's usally returned by
the API call which returns Files. A group of Files is defined as FilesList.
Users do not usually need to construct Files or FilesList manually,
they are generated from a API call most of the time.
}
\section{Fields}{

\describe{
\item{\code{id}}{character used as file id}

\item{\code{name}}{string used as file name}

\item{\code{metdata}}{a list for metdata associated with the file}

\item{\code{project}}{project id if any, when returned by a API call,
it usually return the project id and stored with the object.}

\item{\code{url}}{url}

\item{\code{created_on}}{date created on}

\item{\code{modifield_on}}{data modified on}

\item{\code{origin}}{list as origin}
}}
\section{Methods}{

\describe{
\item{\code{add_tag(x, ...)}}{add new tags while keeping old tags}

\item{\code{copy_to(project = NULL, name = NULL)}}{copy a file to a project (id) with new name }

\item{\code{download(destfile, ..., method = "curl")}}{see help(download.file) for more options}

\item{\code{meta()}}{get metadata from a file}

\item{\code{setMeta(..., overwrite = FALSE)}}{set metadata with provided list, when overwrite
is set to TRUE, it overwrites the metadata}

\item{\code{set_meta(..., overwrite = FALSE)}}{set metadata with provided list, when overwrite
is set to TRUE, it overwrites the metadata}

\item{\code{set_tag(x = NULL, overwrite = TRUE, ...)}}{set a tag for a file, your tag need to be a list or vector}

\item{\code{tag()}}{get tag from a file}

\item{\code{update(name = NULL, metadata = NULL, tags = NULL)}}{This call updates the name, the full set metadata,
and tags for a specified file.}
}}
\examples{
Files(id = "test_id", name = "test.bam")
}

