% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS-class.R
\docType{class}
\name{FS-class}
\alias{FS}
\alias{FS-class}
\title{FS class}
\arguments{
\item{serverAddress}{placehoder}

\item{apiAddress}{placehoder}

\item{vsfsJar}{placehoder}

\item{cacheDir}{placehoder}

\item{cacheSize}{placeholder}

\item{projectID}{placeholder}
}
\description{
FS class
}
\section{Methods}{

\describe{
\item{\code{files(id = NULL)}}{given project id, show all files in it}

\item{\code{mount(mountPoint = NULL, projectId = NULL, ignore.stdout = TRUE,
  sudo = TRUE, ...)}}{mount a specific project if projectId is provided, otherwise mount all projects}

\item{\code{path(id = NULL)}}{list path for all mounted projects, for easy copy/paste of file path; if project id
is provoded, show project path and files path}

\item{\code{unmount(mountCmd = NULL, projectID = NULL, ...)}}{unmount a project if projectID is provided, otherwise unmount all}
}}

