% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-misc.R
\name{get_token}
\alias{get_token}
\alias{misc_get_token}
\title{Opens web browser to copy the auth token}
\usage{
get_token(platform = c("cgc", "aws-us", "aws-eu", "gcp", "cavatica"))

misc_get_token()
}
\arguments{
\item{platform}{The Seven Bridges platform to use.}
}
\value{
auth token
}
\description{
Click the "Generate Token" or "Regenerate" button, copy and paste
the authentication token string to the R console.
The function will return the token string.
}
\examples{
token = NULL
# Will be prompted to enter the auth token
\dontrun{token = get_token(platform = "cgc")}
}
