% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auth-class.R
\name{getToken}
\alias{getToken}
\title{get Token}
\usage{
getToken(platform = NULL, username = NULL)
}
\arguments{
\item{platform}{In your configure file, platform you want to access via API.}

\item{username}{username to specify the token associated with.}
}
\description{
get Token from config files and option list
}
\details{
Current config file is set on home directory with the name .sbg.auth.yml
}

