% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-handler.R
\docType{class}
\name{Handler-class}
\alias{Handler}
\alias{Handler-class}
\title{Handler instance}
\value{
a Handler object
}
\description{
Create Handler instance
}
\details{
Used for parse R Markdown and lift into command line interface,
Dockerfile, Docker container, and cwl json.
}
\section{Fields}{

\describe{
\item{\code{dockerfileHandler}}{a function or NULL, how you handle Dockefile,
for example, push it to GitHub.}

\item{\code{dockerHandler}}{a function or NULL, how you handle local docker
container, for example, push it to DockerHub.}

\item{\code{cwlHandler}}{a function or NULL, how you handle cwl json file or
yaml file, for example, push it to SevenBridges platform as an app.}
}}

