% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{input_matrix}
\alias{input_matrix}
\alias{output_matrix}
\title{get input/output matrix out of json cwl file directly}
\usage{
input_matrix(from, new.order = c("id", "label", "type", "required", "prefix",
  "fileTypes"), required = NULL)

output_matrix(from, new.order = c("id", "label", "type", "fileTypes"))
}
\arguments{
\item{from}{a json file path}

\item{new.order}{a vector of column orders by default for input it's
"id", "label", "type", "required", "prefix", "fileTypes"; For output it's
 "id", "label", "type", "fileTypes"}

\item{required}{logical value, show requried input node only or not.}
}
\description{
an efficient way to access json file, no need to convert a json into a Tool or Flow object before access, directly
operate on a list parsed from json file. Compare to \code{convert_app}, it's much faster.
}
\examples{
tool.in = system.file("extdata/app", "tool_unpack_fastq.json", package = "sevenbridges")
flow.in = system.file("extdata/app", "flow_star.json", package = "sevenbridges")
input_matrix(tool.in)
input_matrix(tool.in, required = TRUE)
input_matrix(flow.in)
input_matrix(flow.in, c("id", "type"))
input_matrix(flow.in, required = TRUE)
tool.in = system.file("extdata/app", "tool_unpack_fastq.json", package = "sevenbridges")
flow.in = system.file("extdata/app", "flow_star.json", package = "sevenbridges")
output_matrix(tool.in)
output_matrix(flow.in)
}

