% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-item.R
\docType{class}
\name{Item-class}
\alias{Item}
\alias{Item-class}
\title{Class Item}
\description{
Class Item
}
\details{
Base class for describing a set of objects:
Project, Task, Pipeline, Files, etc.
}
\section{Fields}{

\describe{
\item{\code{response}}{save the raw response from a request.}

\item{\code{auth_token}}{propagate the auth_token from parent.}

\item{\code{href}}{API href}
}}

