% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-misc.R
\name{misc_get_uploader}
\alias{misc_get_uploader}
\title{Download SBG uploader and extract to a specified directory}
\usage{
misc_get_uploader(destdir = NULL)
}
\arguments{
\item{destdir}{The directory to extract SBG uploader to.
If not present, it will be created automatically.}
}
\value{
\code{0L} if the SBG CLI uploader is successfully
downloaded and unarchived.
}
\description{
Download SBG uploader and extract to a specified directory.
}
\examples{
dir = "~/sbg-uploader/"
\donttest{misc_get_uploader(dir)}
}

