% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pipeline_details}
\alias{pipeline_details}
\title{Returns the details of a pipeline for a project}
\usage{
pipeline_details(auth_token = NULL, project_id = NULL, pipeline_id = NULL,
  ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{pipeline_id}{ID of a pipeline you want to access.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns the details of a pipeline (runtime and regular parameters,
description etc.) for a project.
}
\details{
When using the API to run a task, the user needs to set input files
for all input nodes. To facilitate this, some pipeline input nodes
may contain field "suggested files", that contains files which may
be used as default input (reference genomes, SNP database, etc.).
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = pipeline_details(token,
            project_id = 'b0b3a611-6bb0-47e5-add7-a83402cf7858',
            pipeline_id = '55606ad4896a5d524656afd0')}
}

