% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pipeline_list_my}
\alias{pipeline_list_my}
\title{Returns the list of pipelines in user's "My Pipelines" section}
\usage{
pipeline_list_my(auth_token = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns the list of pipelines in user's "My Pipelines" section.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = pipeline_list_my(token)}
}

