% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pipeline_list_project}
\alias{pipeline_list_project}
\title{Returns a list of all the pipelines in project}
\usage{
pipeline_list_project(auth_token = NULL, project_id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns a list of all the pipelines in project.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = pipeline_list_project(token,
            project_id = 'b0b3a611-6bb0-47e5-add7-a83402cf7858')}
}

