% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pipeline_list_pub}
\alias{pipeline_list_pub}
\title{Returns the list of all public pipelines}
\usage{
pipeline_list_pub(auth_token = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns the list of all public pipelines.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = pipeline_list_pub(token)}
}

