% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{project_delete}
\alias{project_delete}
\title{Delete a project}
\usage{
project_delete(auth_token = NULL, project_id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to delete.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Note that this deletes all files, tasks which belong to a project.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = project_delete(token,
            project_id = '3a21ade8-ef3e-41f8-8ac2-1dc3b434ac77')}
}

