% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{project_details}
\alias{project_details}
\title{Returns the details of the project}
\usage{
project_details(auth_token = NULL, project_id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns the details of the project.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = project_details(token,
                project_id = 'b0b3a611-6bb0-47e5-add7-a83402cf7858')}
}

