% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{project_member_add}
\alias{project_member_add}
\title{Add a user to the project with appropriate permissions}
\usage{
project_member_add(auth_token = NULL, project_id = NULL, username = NULL,
  copy = FALSE, write = FALSE, execute = FALSE, admin = FALSE, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{Name of the project you wish to add user to.}

\item{username}{SBG platform username for a user you wish to add to
the project.}

\item{copy}{Logical. Ability to download or copy files.}

\item{write}{Logical. Ability to create/edit/delete project objects.}

\item{execute}{Logical. Ability to run tasks.}

\item{admin}{Logical. User has all rights on the project
(including changing).}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
You can use this call to add specific users to a project and set their
privileges. Note that you need to specify user's SBG platform username
when adding to the project.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = project_member_add(token,
                project_id = '88fc89c1-cfcd-46ed-a830-6a2fc110c628',
                username = 'testuser', write = TRUE)}
}

