% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{project_member_delete}
\alias{project_member_delete}
\title{Removes a member from a project}
\usage{
project_member_delete(auth_token = NULL, project_id = NULL,
  user_id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{user_id}{ID of the user you wish to remove.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Note that user_id parameter is not username, but user ID parameter
that you can receive from GET members call.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = project_member_delete(token,
            project_id = '7f7a72d0-da77-4f51-9416-99f14f7316ab',
            user_id = '08890148-6d9e-4a10-b284-924228d3f99a')}
}

