% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{project_member_update}
\alias{project_member_update}
\title{Set permissions for a user to a project}
\usage{
project_member_update(auth_token = NULL, project_id = NULL,
  user_id = NULL, write = FALSE, copy = FALSE, execute = FALSE,
  admin = FALSE, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{user_id}{ID of a user whose permissions you with to set}

\item{write}{Logical. Ability to create/edit/delete project objects.}

\item{copy}{Logical. Ability to download or copy files.}

\item{execute}{Logical. Ability to run tasks.}

\item{admin}{Logical. User has all rights on the project
(including changing).}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
This call will set project's member privileges.
Privileges you do not explicitly set to "true" will be automatically
set to "false". Project ID and user ID are specified in path parameters.
Note that you must get the user IDs by performing the project_members()
call and gathering id of the user with a specific permission.
}
\examples{
token = '58aeb140-1970-0130-6386-001f5b34aa78'
\donttest{req = project_member_update(token,
            project_id = '7f7a72d0-da77-4f51-9416-99f14f7316ab',
            user_id = '08890148-6d9e-4a10-b284-924228d3f99a')}
}

