% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{project_new}
\alias{project_new}
\title{Create new project}
\usage{
project_new(auth_token = NULL, name = NULL, description = NULL,
  billing_group_id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{name}{Name of the project you wish to create.}

\item{description}{Description of the project you wish to create.}

\item{billing_group_id}{ID of the billing group you wish to use
for this project.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
You can use this call to create a project. All details, including
project name, description and funding source are specified as part
of the JSON, sent as the body of the request. This call returns
details of the project.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = project_new(token, name = 'Test API project',
                description = 'My first API project',
                billing_group_id = '5b6d5e71-dff8-42fc-8583-500d858f1093')}
}

