% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tool-class.R
\docType{class}
\name{CPURequirement-class}
\alias{CPURequirement}
\alias{CPURequirement-class}
\alias{MemRequirement}
\alias{MemRequirement-class}
\title{Rabix specifc Requirements}
\description{
Extends ProcessRequirements. CPURequirement and MemRequirement to
setup CPU and Memory requiremnts.
}
\section{Fields}{

\describe{
\item{\code{value}}{[Integer] for CPU default is 1L, if 0L, use all
CPU. For mem, default is 1000L. Note: for CPU, 0L means
multi-tread, and non-zero value will be converted to 1L, which
means single thread.}
}}
\examples{
CPURequirement(value = 1L)
MemRequirement(value = 2000L)
}

