% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-auth-utils.R
\name{sbg_set_env}
\alias{sbg_set_env}
\title{Set authentication environment variables for Seven Bridges API}
\usage{
sbg_set_env(url = NULL, token = NULL)
}
\arguments{
\item{url}{Base URL for API.}

\item{token}{Your authentication token.}
}
\value{
set two environment variables
}
\description{
Set authentication environment variables for Seven Bridges API
}
\examples{
# set and get two environment variables for Cancer Genomics Cloud
\dontrun{
sbg_set_env("https://cgc-api.sbgenomics.com/v2",
            "aef7e9f3fec54fb1b438ac4fcddf1a56")
sbg_get_env("SB_API_ENDPOINT")
sbg_get_env("SB_AUTH_TOKEN")}
}

