% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgr-package.R
\docType{package}
\name{sbgr-package}
\alias{sbgr-package}
\title{R Client for Seven Bridges Genomics API}
\description{
R Client for Seven Bridges Genomics API
}
\details{
The vignette can be opened with \code{vignette('sbgr')}.

\tabular{ll}{ Package: \tab sbgr\cr Type: \tab Package\cr
License: \tab MIT\cr }
}
\author{
Nan Xiao <\email{nan.xiao@sbgenomics.com}>
        Tengfei Yin <\email{tengfei.yin@sbgenomics.com}>
}

