% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{set_test_env}
\alias{set_test_env}
\title{Set testing env}
\usage{
set_test_env(type = "host", docker_image = "tengfei/testenv",
  data_dir = getwd())
}
\arguments{
\item{type}{"dind" or "host"}

\item{docker_image}{required docker image with
pre-installed bunny, default: tengfei/testenv}

\item{data_dir}{directory with data to mount
(also will be execution directory)}
}
\value{
docker stdout
}
\description{
Checks if docker is installed, is running and has
required images downloaded and if do creates container
}
\examples{
\dontrun{
set_test_env("dind", "tengfei/testenv", "/Users/<user>/tools")
}
}

