% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{setListClass}
\alias{setListClass}
\title{List Class generator.}
\usage{
setListClass(
  elementType = NULL,
  suffix = "List",
  contains = NULL,
  where = topenv(parent.frame())
)
}
\arguments{
\item{elementType}{[character]}

\item{suffix}{[character] default is "List"}

\item{contains}{[character] class name.}

\item{where}{environment.}
}
\value{
S4 class constructor
}
\description{
Extends IRanges SimpleList class and return constructor.
}
