% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sevenbridges-package.R
\docType{package}
\name{sevenbridges-package}
\alias{sevenbridges-package}
\title{Seven Bridges Platform API Client and CWL Tool Builder in R}
\description{
R client and utilities for Seven Bridges platform API, from Cancer
Genomics Cloud to other Seven Bridges supported platforms.
}
\details{
\tabular{ll}{
Package: \tab sevenbridges\cr
Type: \tab Package\cr
License: \tab Apache License 2.0}
}
\author{
Soner Koc <\email{soner.koc@sevenbridges.com}>
Nan Xiao
Tengfei Yin
Dusan Randjelovic
Emile Young
}
