% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{status_check}
\alias{status_check}
\title{check request status}
\usage{
status_check(req, as = "parsed", ...)
}
\value{
request content or the message
}
\description{
check request status
}
\keyword{internal}

