% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File-class.R
\docType{methods}
\name{set_tag}
\alias{add_tag}
\alias{add_tag,Files-method}
\alias{add_tag,FilesList-method}
\alias{set_tag}
\alias{set_tag,Files-method}
\alias{set_tag,FilesList-method}
\title{set file tags}
\usage{
set_tag(obj, ...)

\S4method{set_tag}{FilesList}(obj, ...)

\S4method{set_tag}{Files}(obj, ...)

add_tag(obj, ...)

\S4method{add_tag}{FilesList}(obj, ...)

\S4method{add_tag}{Files}(obj, ...)
}
\arguments{
\item{obj}{single File or FileList}

\item{...}{passed to obj$set_tag() or obj$add_tag()}
}
\value{
tag list

tag list
}
\description{
set file tags

add new file tags and keep the old tags
}
\examples{
\dontrun{
fl = a$project("demo")$file("omni")
set_tag(fl, "new tag")
set_tag(fl, list("new tag", "new tag 2"))
}
\dontrun{
fl = a$project("demo")$file("omni")
add_tag(fl, "new tag")
add_tag(fl, list("new tag", "new tag 2"))
}
}

