% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{task_action}
\alias{task_action}
\title{Performs action on the task}
\usage{
task_action(auth_token = NULL, project_id = NULL, task_id = NULL,
  action = "abort", ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{task_id}{ID of a task you want to access.}

\item{action}{Character string specifying the action.
Currently, only supported action is \code{'abort'}.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Performs action on the task.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = task_action(token,
                project_id = '7f7a72d0-da77-4f51-9416-99f14f7316ab',
                task_id = '5506a44ae4b04a4ab3ae7250',
                action = 'abort')}
}

