% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{task_details}
\alias{task_details}
\title{Returns information about the task}
\usage{
task_details(auth_token = NULL, project_id = NULL, task_id = NULL,
  download.url = FALSE, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{task_id}{ID of a task you want to access.}

\item{download.url}{Logical. Return the download URL or not.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns information about the task.
}
\details{
Each task has a status and status message, containing the more detailed
information about the task status, associated with it. This is a list of
all values that task status can have:
\itemize{
\item \code{active} - Task is currently running.
\item \code{completed} - Task has finished successfully.
\item \code{aborted} - Task was aborted by user.
\item \code{failed} - Task has failed to finish due to
either bad inputs and/or parameters, or because of the
internal infrastructure failures.}
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req1 = task_details(token,
                 project_id = '1c1d06d2-5862-48f6-b595-e0099b20937e',
                 task_id = '22237')
req2 = task_details(token,
       project_id = '1c1d06d2-5862-48f6-b595-e0099b20937e',
       task_id = '22237', download.url = TRUE)}
}

