% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{task_run}
\alias{task_run}
\title{Runs a task as a part of a project}
\usage{
task_run(auth_token = NULL, project_id = NULL, task_details = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{task_details}{A list with the following components:
\itemize{
\item \code{pipeline_id} - ID of the pipeline you wish to execute
\item \code{pipeline_revision} - Revision of the pipeline you wish
to execute. If not specified, latest revision is used.
\item \code{name} - Name of the task you wish to execute
\item \code{description} - Description of the task you wish to execute
\item \code{inputs} - Named list containing mappings of pipeline input
node ID to file IDs. Note that file IDs always need to be specified as
an list, even if empty or with one element.
\item \code{parameters} - Named list containing mappings of node IDs
to apps specific parameters. Note that parameters are always specified
as an list, even if empty or with one element.}}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Runs a task as a part of a project.
}
\details{
All the details, including the pipeline ID and runtime parameters,
are specified via a list. See the example for details.
}
\examples{
token = '58aeb140-1970-0130-6386-001f5b34aa78'
details = list(
  'name' = 'Test 2 of C. Elegans VC',
  'description' = 'Testing Caenorhabditis elegans Exome Variant Calling',
  'pipeline_id' = '422',
  'inputs' = list('309485' = 13645,
                  '317344' = 13646,
                  '318662' = 13645,
                  '699018' = 13647),
  'parameters' = list('393463' = list('read_trimming_qual' = 30,
                                      'rg_seq_tech' = 'Illumina'),
                      '677492' = list()))

\donttest{req = task_run(token,
                project_id = '7f7a72d0-da77-4f51-9416-99f14f7316ab',
                task_details = details)}
}

