% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task-class.R
\name{setTaskHook}
\alias{getTaskHook}
\alias{setTaskHook}
\title{set task function hook}
\usage{
setTaskHook(status = c("queued", "draft", "running", "completed", "aborted",
  "failed"), fun)

getTaskHook(status = c("queued", "draft", "running", "completed", "aborted",
  "failed"))
}
\arguments{
\item{status}{one of ("queued", "draft", "running", "completed", "aborted", "failed")}

\item{fun}{function}
}
\value{
object from setHook and getHook.
}
\description{
set task function hoook according to
}
\examples{
getTaskHook("completed")
setHook("completed", function(){
    message("completed yet ...")
})
}

