% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_tool}
\alias{test_tool}
\title{Test tools in rabix}
\usage{
test_tool(rabix_tool, inputs)
}
\arguments{
\item{rabix_tool}{rabix tool from Tool class}

\item{inputs}{input parameters declared as json (or yaml) string}
}
\value{
bunny stdout
}
\description{
Test tools locally in rabix/bunny inside docker container
}
\examples{
\dontrun{
inputs <- '{"counts_file": {"class": "File", "path": "./FPKM.txt"}, "gene_names": "BRCA1"}'
test_tool(bunny, write(rbx$toJSON, file="/data_dir/tool.json"), write(inputs, file="/data_dir/inputs.json"))
}
}

