% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_tool_bunny}
\alias{test_tool_bunny}
\title{Test tools in rabix/bunny}
\usage{
test_tool_bunny(rabix_tool, inputs)
}
\arguments{
\item{rabix_tool}{rabix tool from Tool class}

\item{inputs}{input parameters declared as json (or yaml) string}
}
\value{
bunny stdout
}
\description{
Test tools locally in rabix/bunny inside docker container
}
\examples{
\dontrun{
inputs <- '{"counts_file": {"class": "File", "path": "./FPKM.txt"}, "gene_names": "BRCA1"}'
rbx <- <define rabix tool>
set_test_env("tengfei/testenv", "<mount_dir>")
test_tool_bunny(rbx, inputs)
}
}

