% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_tool_cwlrun}
\alias{test_tool_cwlrun}
\title{Test tools with cwl-runner}
\usage{
test_tool_cwlrun(rabix_tool, inputs = list())
}
\arguments{
\item{rabix_tool}{rabix tool from Tool class}

\item{inputs}{input parameters declared as json (or yaml) string}
}
\value{
cwl-runner stdout
}
\description{
Test tools locally cwl-runner (https://github.com/common-workflow-language/cwltool)
}
\examples{
\dontrun{
inputs <- '{"counts_file": {"class": "File", "path": "./FPKM.txt"}, "gene_names": "BRCA1"}'
rbx <- <define rabix tool>
set_test_env("tengfei/testenv", "<mount_dir>")
test_tool_cwlrun(rbx, inputs)
}
}

