% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{test_tool_rabix}
\alias{test_tool_rabix}
\title{Test tools in rabix/rabix-devel (DEPRECATED)}
\usage{
test_tool_rabix(rabix_tool, inputs = list())
}
\arguments{
\item{rabix_tool}{rabix tool from Tool class}

\item{inputs}{input parameters declared as json (or yaml) string}
}
\value{
rabix stdout
}
\description{
Test tools locally in rabix/rabix-devel python executor (DEPRECATED)
}
\examples{
\dontrun{
inputs <- '{"counts_file": {"class": "File", "path": "./FPKM.txt"}, "gene_names": "BRCA1"}'
rbx <- <define rabix tool>
set_test_env("tengfei/testenv", "<mount_dir>")
test_tool_rabix(rbx, inputs)}
}

