% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-http.R
\name{upload_delete}
\alias{upload_delete}
\title{Aborts the upload}
\usage{
upload_delete(token = NULL, upload_id = NULL, ...)
}
\arguments{
\item{token}{auth token}

\item{upload_id}{ID of the upload}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
All upload records and the file are deleted.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = upload_delete(token,
            upload_id = '8D7sQJxQk14ubsEnKaoeQZlRvV6ouQtMzBWaQNJdxPDLypUC3WogwtJdncevHxnT')}
}

