% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App-class.R
\docType{class}
\name{App-class}
\alias{App}
\alias{App-class}
\alias{appType}
\title{App class}
\usage{
appType(x)
}
\arguments{
\item{x}{a App object}
}
\value{
App object.
}
\description{
App class
}
\section{Fields}{

\describe{
\item{\code{id}}{app id.}

\item{\code{project}}{project id}

\item{\code{name}}{app name}

\item{\code{revision}}{app revision}

\item{\code{raw}}{raw cwl list, if doesn't have any, call cwl() method.}
}}
\section{appType}{

\describe{
 this function return class of a App object.
}
}

