% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File-class.R
\docType{methods}
\name{download}
\alias{download}
\title{Download file or files}
\usage{
download(obj, ...)
}
\arguments{
\item{obj}{single File or FileList}

\item{...}{passed to download()}
}
\description{
Download file
}
\examples{
\dontrun{
a$project("demo")$file("omni")$download()
## or
download(a$project("demo")$file("omni"))
}
}

