% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File-class.R
\docType{class}
\name{Files-class}
\alias{Files}
\alias{Files-class}
\title{Files class}
\value{
Files object
}
\description{
Files Class
}
\details{
Files class (has 's') contains CWL File class, it's usally returned by the API call which 
returns Files. A group of Files is defined as FilesList. Users don't usually
need to construct Files or FilesList manually, most time it's generated from a API 
call.
}
\section{Fields}{

\describe{
\item{\code{id}}{character used as file id}

\item{\code{name}}{string used as file name}

\item{\code{metdata}}{a list for metdata associated with the file}

\item{\code{project}}{project id if any, when returned by a API call, it usually return the 
project id and stored with the object.}

\item{\code{url}}{url}

\item{\code{created_on}}{date created on}

\item{\code{modifield_on}}{data modified on}

\item{\code{origin}}{list as origin}
}}
\section{Methods}{

\describe{
\item{\code{download(destfile, ..., method = "curl")}}{see help(download.file) for more options}

\item{\code{meta()}}{get metadata from a file}
}}
\examples{
Files(id = "test_id", name = "test.bam")
}

