% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow-class.R
\docType{class}
\name{SBGWorkflow-class}
\alias{"+",App,App-method}
\alias{"+",Tool,Tool-method}
\alias{"+",WorkflowStepList,App-method}
\alias{"+",WorkflowStepList,Tool-method}
\alias{"+",WorkflowStepList,WorkflowStepList-method}
\alias{"\%>\%"}
\alias{"\%>\%",App,App-method}
\alias{"\%>\%",Tool,Tool-method}
\alias{"\%>\%",Workflow,App-method}
\alias{"\%>\%",Workflow,Tool-method}
\alias{"\%>\%",Workflow,Workflow-method}
\alias{+,App,App-method}
\alias{+,Tool,Tool-method}
\alias{+,WorkflowStepList,App-method}
\alias{+,WorkflowStepList,Tool-method}
\alias{+,WorkflowStepList,WorkflowStepList-method}
\alias{Flow}
\alias{SBGWorkflow}
\alias{SBGWorkflow-class}
\alias{\%>\%}
\alias{\%>\%,App,App-method}
\alias{\%>\%,Tool,Tool-method}
\alias{\%>\%,Workflow,App-method}
\alias{\%>\%,Workflow,Tool-method}
\alias{\%>\%,Workflow,Workflow-method}
\title{Build workflow}
\usage{
Flow(..., graph = TRUE, x.width = 1000, y.width = 400, x.start = 100,
  y.start = 200, canvas_zoom = 1, canvas_x = 40, canvas_y = 130)

\S4method{+}{Tool,Tool}(e1, e2)

\S4method{+}{WorkflowStepList,Tool}(e1, e2)

\S4method{+}{WorkflowStepList,WorkflowStepList}(e1, e2)

\S4method{+}{App,App}(e1, e2)

\S4method{+}{WorkflowStepList,App}(e1, e2)

e1 \%>\% e2

\S4method{\%>\%}{Tool,Tool}(e1, e2)

\S4method{\%>\%}{Workflow,Tool}(e1, e2)

\S4method{\%>\%}{Workflow,Workflow}(e1, e2)

\S4method{\%>\%}{App,App}(e1, e2)

\S4method{\%>\%}{Workflow,App}(e1, e2)
}
\arguments{
\item{...}{extra arguments passed to SBGWorkflow}

\item{graph}{if add graph coordinates or not, used for visualization on seven bridges platform.}

\item{x.width}{x scale width}

\item{y.width}{y scale width}

\item{x.start}{node x start point for a flow}

\item{y.start}{node y start point for a flow}

\item{canvas_zoom}{zoom factor}

\item{canvas_x}{canvas x}

\item{canvas_y}{canvas y}

\item{e1}{either Tool App or Workflow object}

\item{e2}{either Tool App or Workflow object}
}
\value{
a SBGWorkflow object.
}
\description{
Build workflow
}

