% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta-class.R
\docType{class}
\name{Metadata-class}
\alias{Metadata}
\alias{Metadata-class}
\alias{access_level}
\alias{age_at_diagnosis}
\alias{aliquot_id}
\alias{aliquot_uuid}
\alias{analysis_uuid}
\alias{case_id}
\alias{case_uuid}
\alias{data_format}
\alias{data_subtype}
\alias{data_type}
\alias{days_to_death}
\alias{disease_type}
\alias{ethnicity}
\alias{experimental_strategy}
\alias{file_extension}
\alias{file_segment_number}
\alias{gdc_file_uuid}
\alias{gender}
\alias{investigation}
\alias{library_id}
\alias{paired_end}
\alias{platform}
\alias{platform_unit_id}
\alias{primary_site}
\alias{quality_scale}
\alias{race}
\alias{reference_genome}
\alias{sample_id}
\alias{sample_type}
\alias{sample_uuid}
\alias{vital_status}
\title{Meta schema}
\description{
Meta schema
}
\details{
V2 version for meta data schema
}

