% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-meta.R
\docType{class}
\name{Metadata-class}
\alias{Metadata}
\alias{Metadata-class}
\alias{access_level}
\alias{age_at_diagnosis}
\alias{aliquot_id}
\alias{aliquot_uuid}
\alias{analysis_uuid}
\alias{case_id}
\alias{case_uuid}
\alias{data_format}
\alias{data_subtype}
\alias{data_type}
\alias{days_to_death}
\alias{disease_type}
\alias{ethnicity}
\alias{experimental_strategy}
\alias{file_extension}
\alias{file_segment_number}
\alias{gdc_file_uuid}
\alias{gender}
\alias{investigation}
\alias{library_id}
\alias{paired_end}
\alias{platform}
\alias{platform_unit_id}
\alias{primary_site}
\alias{quality_scale}
\alias{race}
\alias{reference_genome}
\alias{sample_id}
\alias{sample_type}
\alias{sample_uuid}
\alias{vital_status}
\title{Meta schema}
\description{
Meta schema
}
\details{
V2 version for meta data schema
}
\examples{
# show schema (you can still provide customized one)
# empty beause they are all NULL
Metadata()
# show schema
Metadata()$show(TRUE)
# or
names(Metadata()$asList(TRUE))
# returned meta field is actually define as function too, direclty
# call them will give you details
platform()
paired_end()
quality_scale()
# check their suggested value and construct your metadata
Metadata(platform  = "Affymetrix SNP Array 6.0", paired_end = 1, quality_scale = "sanger")
}

