% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-misc.R
\name{misc_get_token}
\alias{misc_get_token}
\title{Opens web browser to copy the auth token}
\usage{
misc_get_token()
}
\value{
auth token
}
\description{
Click the "Generate Token" or "Regenerate" button, copy and paste
the authentication token string to the R console.
The function will return the token string.
}
\examples{
# Paste the auth token into R
# console then press enter:
token = NULL
\donttest{token = misc_get_token()}
}

