% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-misc.R
\name{project_details}
\alias{project_details}
\title{Returns the details of the project}
\usage{
project_details(token = NULL, project_id = NULL, ...)
}
\arguments{
\item{token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns the details of the project.
}
\examples{
token = "your_token"
\dontrun{
req = project_details(token, project_id = "your_project_id")}
}
