% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{project_members}
\alias{project_members}
\title{Returns a list of all users invited to the project and their privileges}
\usage{
project_members(token = NULL, project_id = NULL, ...)
}
\arguments{
\item{token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns a list of all users invited to the project and their privileges.
Project ID is specified as path parameter. Call returns ID and username
of the user with privileges.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = project_members(token,
                project_id = 'b0b3a611-6bb0-47e5-add7-a83402cf7858')}
}

