% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-auth-utils.R
\name{sbg_set_env}
\alias{sbg_set_env}
\title{Set authentication environment variables for Seven Bridges API}
\usage{
sbg_set_env(url = NULL, token = NULL)
}
\arguments{
\item{url}{Base URL for API.}

\item{token}{Your authentication token.}
}
\value{
set two environment variables for authentication
}
\description{
Set authentication environment variables for Seven Bridges API
}
\examples{
# set and get environment variables for CGC
token <- "your_token"
\dontrun{
sbg_set_env("https://cgc-api.sbgenomics.com/v2", token)
sbg_get_env("SB_API_ENDPOINT")
sbg_get_env("SB_AUTH_TOKEN")}
}
