% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cwl.R
\docType{class}
\name{SchemaList}
\alias{SchemaList}
\alias{SchemaList-class}
\alias{ScehmaList}
\alias{Schema-class}
\alias{Schema}
\alias{SchemaDef-class}
\alias{SchemaDef}
\alias{SchemaDefList}
\alias{SchemaDefList-class}
\alias{InputSchema-class}
\alias{InputSchema}
\alias{OutputSchema-class}
\alias{OutputSchema}
\title{SchemaList}
\usage{
SchemaList(...)

SchemaDefList(...)
}
\arguments{
\item{\dots}{element or list of the element.}
}
\value{
a Schema object or sbuclass object.
}
\description{
A schema defines a parameter type.
}
\section{Fields}{

\describe{
\item{\code{type}}{(ANY) The data type of this parameter.}

\item{\code{fields}}{[SchemaList] When type is record, defines the fields of the
record.}

\item{\code{symbols}}{[character] When type is enum, defines the set of valid symbols.}

\item{\code{items}}{[ANY] When type is array, defines the type of the array
elements.}

\item{\code{values}}{[ANY] When type is map, defines the value type for the
key/value pairs.}

\item{\code{inputBinding}}{[Binding] Describes how to handle a value in
the input object convert it into a concrete form for execution,
such as command line parameters.}
}}


\examples{
Schema(fields = SchemaList(SchemaDef(name = "schema")))
}
