% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sevenbridges-package.R
\docType{package}
\name{sevenbridges-package}
\alias{sevenbridges-package}
\title{R Client for Seven Bridges Platform API and CWL Tool builder in R}
\description{
R client and utilities for Seven Bridges platform API, from Cancer
Genomics Cloud to other Seven Bridges supported platforms.
}
\details{
\tabular{ll}{ Package: \tab sevenbridges\cr Type: \tab Package\cr
License: \tab Apache License 2.0\cr }
}
\author{
Tengfei Yin <\email{tengfei.yin@sbgenomics.com}>
        Dusan Randjelovic <\email{dusan.randjelovic@sbgenomics.com}>
        Nan Xiao <\email{nan.xiao@sbgenomics.com}>
}

