% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sevenbridges-package.R
\docType{package}
\name{sevenbridges-package}
\alias{sevenbridges-package}
\title{R Client for Seven Bridges Genomics API}
\description{
R Client for Seven Bridges Genomics API
}
\details{
The vignette can be opened with \code{vignette('sevenbridges')}.

\tabular{ll}{ Package: \tab sevenbridges\cr Type: \tab Package\cr
License: \tab MIT\cr }
}
\author{
Nan Xiao <\email{nan.xiao@sbgenomics.com}>
        Tengfei Yin <\email{tengfei.yin@sbgenomics.com}>
}

