% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-http.R
\name{status_check}
\alias{status_check}
\title{Check request status}
\usage{
status_check(req, as = "parsed", ...)
}
\value{
request content or the message
}
\description{
Check request status
}
\keyword{internal}
