% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App-class.R, R/Tool-class.R
\docType{class}
\name{convertApp}
\alias{Tool}
\alias{Tool-class}
\alias{convertApp}
\title{Rabix CommandLineTool Class}
\usage{
convertApp(from)
}
\arguments{
\item{from}{an App object}
}
\value{
a Tool object.
}
\description{
Rabix subclass for CommandLineTool used by rabix.org or sbg
platform. \code{Tool} class extends \code{CommandLineTool}
with more fields.
}
\section{Fields}{

\describe{
\item{\code{context}}{[character] by default:
"https://github.com/common-workflow-language/common-workflow-language/blob/draft-1/specification/tool-description.md"}

\item{\code{owner}}{[list] a list of owner names.}

\item{\code{contributor}}{[list] a list of contributor names.}
}}
\section{convertApp}{

\describe{
convert App into Tool or Flow object based on class. 
}
}

