% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tool-class.R
\docType{class}
\name{Tool-class}
\alias{Tool}
\alias{Tool-class}
\title{Rabix CommandLineTool Class}
\value{
a Tool object.
}
\description{
Rabix subclass for CommandLineTool used by rabix.org or sbg
platform. \code{Tool} class extends \code{CommandLineTool}
with more seven bridges fields.
}
\section{Fields}{

\describe{
\item{\code{context}}{[character] by default:
"https://github.com/common-workflow-language/common-workflow-language/blob/draft-1/specification/tool-description.md"}

\item{\code{owner}}{[list] a list of owner names.}

\item{\code{contributor}}{[list] a list of contributor names.}
}}

