% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-misc.R
\name{upload_complete_all}
\alias{upload_complete_all}
\title{Reports the complete file upload}
\usage{
upload_complete_all(token = NULL, upload_id = NULL, ...)
}
\arguments{
\item{token}{auth token}

\item{upload_id}{ID of the upload}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
If the whole parts are uploaded, and the provided ETags are correct,
then the file is assembled and made available on the SBG platform.
}
\examples{
token = "your_token"
\dontrun{
req = upload_complete_all(token, upload_id = "your_upload_id")}
}
