% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-misc.R
\name{upload_delete}
\alias{upload_delete}
\title{Abort the upload}
\usage{
upload_delete(token = NULL, upload_id = NULL, ...)
}
\arguments{
\item{token}{auth token}

\item{upload_id}{ID of the upload}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Abort the upload; all upload records and the file are deleted.
}
\examples{
token = "your_token"
\donttest{
req = upload_delete(token,
      upload_id = "8D7sQJxQk14ubsEnKaoeQZlRvV6ouQtMzBWaQNJdxPDLypUC3WogwtJdncevHxnT")}
}
