% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-http.R
\name{upload_info}
\alias{upload_info}
\title{Returns upload information for the ongoing upload}
\usage{
upload_info(token = NULL, upload_id = NULL, ...)
}
\arguments{
\item{token}{auth token}

\item{upload_id}{ID of the upload}

\item{...}{parameters passed to api function}
}
\value{
parsed list of the returned json
}
\description{
Returns the upload information for the ongoing upload.
}
\examples{
token = "your_token"
\donttest{req = upload_info(token,
                upload_id = "8D7sQJxQk14ubsEnKaoeQZlRvV6ouQtMzBWaQNJdxPDLypUC3WogwtJdncevHxnT")}
}

